requirements.txt
pandas==1.3.5
pyyaml==6.0
numpy==1.22.1
numpy_financial==1.0.0